Implement proof-of-concept server
At this stage, the server just accepts an RSS or Atom feed in a HTTP
POST request, parses it, and sends webmentions the links in each post
in the feed.  It does not handle errors, it does not perform mentions
asynchronously, it does not handle WebSub subscription, nor does it
insist on an HMAC from the hub.

License: Apache2.0